Refactoring back the original comentarioService.js file. This is done temporarily to maintain functionallity for the other view at VistaLugar.jsx.
in this commit, both methods should work without conflicts, as they point to different tables in the db.
